Task-14: Admin users and order section 1. Add order section for admin where admin can manage orders 2. Add user section where admin can manage users

+++ app/models/user.rb
@@ -37,6 +37,7 @@ class User < ApplicationRecord
   end
 
   def destroy_authy_user
+    return if self.authy_id == 0
     response = Authy::API.delete_user(:id => self.authy_id)
     unless response.ok?
       fail response.errors
